+++ app/helpers/additionals_settings_helper.rb
@@ -14,6 +14,9 @@ module AdditionalsSettingsHelper
   end
 
   def additionals_settings_checkbox(name, options = {})
+    active_value = options.delete(:active_value).presence || @settings.present? && @settings[name]
+    tag_name = options.delete(:tag_name).presence || "settings[#{name}]"
+
     label_title = options.delete(:label).presence || l("label_#{name}")
     value = options.delete :value
     value_is_bool = options.delete :value_is_bool
@@ -26,23 +29,41 @@ module AdditionalsSettingsHelper
                    end
 
     checked = if custom_value && !value_is_bool
-                @settings[name]
+                active_value
               else
-                Additionals.true? @settings[name]
+                Additionals.true? active_value
               end
 
-    s = [label_tag("settings[#{name}]", label_title)]
-    s << hidden_field_tag("settings[#{name}]", 0, id: nil) if !custom_value || value_is_bool
-    s << check_box_tag("settings[#{name}]", value, checked, options)
+    s = [label_tag(tag_name, label_title)]
+    s << hidden_field_tag(tag_name, 0, id: nil) if !custom_value || value_is_bool
+    s << check_box_tag(tag_name, value, checked, options)
     safe_join s
   end
 
-  def additionals_settings_textfield(name, options = {})
-    label_title = options.delete(:label).presence || l("label_#{name}")
-    value = options.delete(:value).presence || @settings[name]
+  def additionals_settings_numberfield(name, options = {})
+    additionals_settings_input_field :number_field_tag, name, options
+  end
 
-    safe_join [label_tag("settings[#{name}]", label_title),
-               text_field_tag("settings[#{name}]", value, options)]
+  def additionals_settings_textfield(name, options = {})
+    additionals_settings_input_field :text_field_tag, name, options
+  end
+
+  def additionals_settings_passwordfield(name, options = {})
+    additionals_settings_input_field :password_field_tag, name, options
+  end
+
+  def additionals_settings_urlfield(name, options = {})
+    additionals_settings_input_field :url_field_tag, name, options
+  end
+
+  def additionals_settings_select(name, values, options = {})
+    tag_name = options.delete(:tag_name).presence || "settings[#{name}]"
+
+    label_title = [options.delete(:label).presence || l("label_#{name}")]
+    label_title << tag.span('*', class: 'required') if options[:required].present?
+
+    safe_join [label_tag(tag_name, safe_join(label_title, ' ')),
+               select_tag(tag_name, values, options)]
   end
 
   def additionals_settings_textarea(name, options = {})
@@ -54,5 +75,22 @@ module AdditionalsSettingsHelper
 
     safe_join [label_tag("settings[#{name}]", label_title),
                text_area_tag("settings[#{name}]", value, options)]
+  end
+
+  private
+
+  def additionals_settings_input_field(tag_field, name, options = {})
+    tag_name = options.delete(:tag_name).presence || "settings[#{name}]"
+    value = if options.key? :value
+              options.delete(:value).presence
+            elsif @settings.present?
+              @settings[name]
+            end
+
+    label_title = [options.delete(:label).presence || l("label_#{name}")]
+    label_title << tag.span('*', class: 'required') if options[:required].present?
+
+    safe_join [label_tag(tag_name, safe_join(label_title, ' ')),
+               send(tag_field, tag_name, value, options)], ' '
   end
 end