+++ app/models/additionals_query.rb
@@ -22,6 +22,17 @@ module AdditionalsQuery
     sql.join(' AND ')
   end
 
+  def fix_sql_for_text_field(field, operator, value, table_name = nil, target_field = nil)
+    table_name = queried_table_name if table_name.blank?
+    target_field = field if target_field.blank?
+
+    sql = []
+    sql << "(#{sql_for_field(field, operator, value, table_name, target_field)})"
+    sql << "#{table_name}.#{target_field} != ''" if operator == '*'
+
+    sql.join(' AND ')
+  end
+
   def initialize_ids_filter(options = {})
     if options[:label]
       add_available_filter 'ids', type: :integer, label: options[:label]
@@ -44,12 +55,26 @@ module AdditionalsQuery
     end
   end
 
+  def sql_for_project_identifier_field(field, operator, values)
+    value = values.first
+    values = value.split(',').map(&:strip) if ['=', '!'].include?(operator) && value.include?(',')
+    sql_for_field field, operator, values, Project.table_name, 'identifier'
+  end
+
   def sql_for_project_status_field(field, operator, value)
     sql_for_field field, operator, value, Project.table_name, 'status'
   end
 
+  def initialize_project_identifier_filter
+    return if project
+
+    add_available_filter 'project.identifier',
+                         type: :string,
+                         name: l(:label_attribute_of_project, name: l(:field_identifier))
+  end
+
   def initialize_project_status_filter
-    return if project&.leaf?
+    return if project
 
     add_available_filter 'project.status',
                          type: :list,
@@ -84,7 +109,7 @@ module AdditionalsQuery
 
   def initialize_tags_filter(options = {})
     add_available_filter 'tags', order: options[:position],
-                                 type: :list,
+                                 type: :list_optional,
                                  values: -> { tag_values(project) }
   end
 
@@ -115,7 +140,7 @@ module AdditionalsQuery
 
     add_available_filter 'watcher_id', order: options[:position],
                                        type: :list,
-                                       values: -> { watcher_values }
+                                       values: -> { watcher_values_for_manage_public_queries }
   end
 
   def tag_values(project)
@@ -139,7 +164,9 @@ module AdditionalsQuery
     assigned_to_values
   end
 
-  def watcher_values
+  # watcher_values of query checks view_issue_watchers, this checks manage_public_queries permission
+  # and with users (not groups)
+  def watcher_values_for_manage_public_queries
     watcher_values = [["<< #{l :label_me} >>", 'me']]
     watcher_values += users.collect { |s| [s.name, s.id.to_s] } if User.current.allowed_to?(:manage_public_queries, project, global: true)
     watcher_values
@@ -155,7 +182,7 @@ module AdditionalsQuery
   end
 
   def sql_for_tags_field(field, _operator, value)
-    AdditionalTags.sql_for_tags_field(queried_class, operator_for(field), value)
+    AdditionalTags.sql_for_tags_field queried_class, operator_for(field), value
   end
 
   def sql_for_is_private_field(_field, operator, value)