+++ lib/additionals/patches/auto_completes_controller_patch.rb
@@ -2,8 +2,8 @@ module Additionals
   module Patches
     module AutoCompletesControllerPatch
       def fontawesome
-        icons = AdditionalsFontAwesome.search_for_select(params[:q].to_s.strip,
-                                                         params[:selected].to_s.strip)
+        icons = AdditionalsFontAwesome.search_for_select params[:q].to_s.strip,
+                                                         params[:selected].to_s.strip
         icons.sort! { |x, y| x[:text] <=> y[:text] }
 
         respond_to do |format|