+++ test/unit/helpers/global_helper_test.rb
@@ -26,17 +26,6 @@ class GlobalHelperTest < ActionView::TestCase
     User.current = nil
   end
 
-  def test_system_info
-    text = system_info
-
-    assert_not_equal '', text
-    assert_not_equal 'unknown', text
-  end
-
-  def test_windows_platform
-    assert_nil windows_platform?
-  end
-
   def test_user_with_avatar
     html = user_with_avatar(users(:users_001))
 
@@ -65,7 +54,7 @@ class GlobalHelperTest < ActionView::TestCase
     assert_equal({ issue_id: nil, comment_id: nil },
                  parse_issue_url('', nil))
     assert_equal({ issue_id: nil, comment_id: nil },
-                 parse_issue_url('http://localhost/issues/23', nil))
+                 parse_issue_url('http://localhost/issue/23', nil))
     assert_equal({ issue_id: '23', comment_id: nil },
                  parse_issue_url('http://redmine.local/issues/23', nil))
     assert_equal({ issue_id: '23', comment_id: 2 },