+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 +++++
 
 - d3plus to v2.0.0-alpha.30 support
+- FontAwesome 5.15.2 support
 
 3.0.1
 +++++