+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 +++++
 
 - d3plus to v2.0.0-alpha.30 support
+- Mermaid 8.9.0 support
+- Bug fix for select2 loading without named field
 
 3.0.1
 +++++