+++ CHANGELOG.rst
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+3.0.2
++++++
+
+- d3plus to v2.0.0-alpha.30 support
+- Mermaid 8.9.2 support
+- Bug fix for select2 loading without named field
+- FontAwesome 5.15.3 support
+- D3 6.6.2 support
+- Fix news limit for welcome dashboard block
+- Frensh translation updated, thanks to Brice BEAUMESNIL!
+- clipboard.js updated to v2.0.8
+- Ruby 2.5 is required
+
 3.0.1
 +++++